lab2/requirement.txt
numpy==1.21.4
Pillow
torch
torchvision
keras